Emit watermarks also from ReplayedSourceFunction

+++ src/main/scala/org/mvrs/dspa/functions/TextFileSourceFunction.scala
@@ -12,7 +12,7 @@ class TextFileSourceFunction[OUT](filePath: String,
                                   parse: String => OUT,
                                   extractEventTime: OUT => Long,
                                   speedupFactor: Double,
-                                  maximumDelayMillis: Long,
+                                  maximumDelayMillis: Int,
                                   delay: OUT => Long)
   extends ReplayedSourceFunction[String, OUT](parse, extractEventTime, speedupFactor, maximumDelayMillis, delay) {
 
@@ -23,7 +23,7 @@ class TextFileSourceFunction[OUT](filePath: String,
            parse: String => OUT,
            extractEventTime: OUT => Long,
            speedupFactor: Double = 0,
-           maximumDelayMilliseconds: Long = 0) =
+           maximumDelayMilliseconds: Int = 0) =
     this(filePath, skipFirstLine, parse, extractEventTime, speedupFactor, maximumDelayMilliseconds,
       if (maximumDelayMilliseconds <= 0) (_: OUT) => 0L
       else (_: OUT) => utils.getNormalDelayMillis(rand, maximumDelayMilliseconds))
